# shared/openapi.yaml
openapi: 3.1.0
info:
  title: GUSTO B2B API
  version: 0.2.0
  description: |
    Контракт API B2B-портала «Густо». Заполняется по мере сессий (чек-лист, Часть 4):
    каждая сессия, меняющая эндпоинт, обновляет этот файл в том же PR.

    **Envelope.** Все ответы: `{data, meta, error}`. Успех — `data` заполнен, `error: null`.
    Ошибка — `data: null`, `error: {code, message, details?}`.

    **Каталог кодов ошибок** (`error.code`) фиксируется в `ErrorCode` и дальше только
    расширяется — существующие коды не переименовываются и не меняют смысл.

    **Пагинация** списков: query `page` (0-based, как в Spring Data), `size`;
    ответ — `meta: {page, size, total}` (total — всего элементов).

    **Аутентификация.** Access-токен (JWT, 15 мин) — заголовок `Authorization: Bearer`.
    Refresh-токен (7 дней, rotation + reuse detection) — httpOnly-кука
    `refresh_token` (Secure, SameSite=Strict), выдаётся при login/refresh,
    используется только эндпоинтом `/auth/refresh` и инвалидируется `/auth/logout`.

    **Идемпотентность.** Мутирующие `POST` (заказы, заявки) принимают заголовок
    `Idempotency-Key` (см. `components.parameters.IdempotencyKey`): повтор с тем же
    ключом в течение 24 ч возвращает тот же результат.

    **Health-эндпоинты** `/healthz`, `/readyz` живут вне префикса `/api/v1`
    (per-path `servers`).
servers:
  - url: /api/v1
...
paths:
...
  /manager/companies:
    get:
      tags: [manager]
      summary: Список назначенных компаний (только MANAGER)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Компании, закреплённые за менеджером
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /manager/companies/{id}:
    parameters:
      - $ref: "#/components/parameters/CompanyId"
    get:
      tags: [manager]
      summary: Карточка компании (MANAGER + ownership)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Данные компании
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [manager]
      summary: Редактирование назначенной компании (MANAGER + ownership)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCompanyRequest" }
      responses:
        "200":
          description: Компания обновлена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/catalog/categories:
    get:
      tags: [admin]
      summary: Список всех категорий
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список категорий
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Создание категории
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryRequest" }
      responses:
        "201":
          description: Категория создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/catalog/categories/tree:
    get:
      tags: [admin]
      summary: Дерево активных категорий
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Дерево категорий
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryTreeResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/catalog/categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID категории
        schema: { type: string, format: uuid }
    get:
      tags: [admin]
      summary: Карточка категории
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Данные категории
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [admin]
      summary: Редактирование категории
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryRequest" }
      responses:
        "200":
          description: Категория обновлена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [admin]
      summary: Удаление категории (только если нет дочерних)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Категория удалена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/catalog/brands:
    get:
      tags: [admin]
      summary: Список брендов
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список брендов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Создание бренда
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrandRequest" }
      responses:
        "201":
          description: Бренд создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/catalog/brands/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID бренда
        schema: { type: string, format: uuid }
    get:
      tags: [admin]
      summary: Карточка бренда
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Данные бренда
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [admin]
      summary: Редактирование бренда
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrandRequest" }
      responses:
        "200":
          description: Бренд обновлен
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [admin]
      summary: Удаление бренда
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Бренд удален
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/catalog/products:
    get:
      tags: [admin]
      summary: Список товаров
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: search
          in: query
          schema: { type: string }
          description: Поиск по части названия
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: brandId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Список товаров с пагинацией
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Создание товара
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductRequest" }
      responses:
        "201":
          description: Товар создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/catalog/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID товара
        schema: { type: string, format: uuid }
    get:
      tags: [admin]
      summary: Карточка товара
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Данные товара
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [admin]
      summary: Редактирование товара
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductRequest" }
      responses:
        "200":
          description: Товар обновлен
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [admin]
      summary: Мягкое удаление товара
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Товар удален
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /catalog/products:
    get:
      tags: [catalog]
      summary: Публичный каталог товаров
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: search
          in: query
          schema: { type: string }
          description: Поиск по части названия
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: brandId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Список товаров с розничными ценами
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogProductListResponse" }
  /catalog/products/{sku}:
    parameters:
      - name: sku
        in: path
        required: true
        description: Артикул товара
        schema: { type: string }
    get:
      tags: [catalog]
      summary: Карточка товара
      responses:
        "200":
          description: Данные товара с розничной ценой
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogProductResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен из /auth/login или /auth/refresh (15 мин)
  parameters:
...
    CompanyId:
      name: id
      in: path
      required: true
      description: ID компании
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Номер страницы, 0-based
    SizeParam:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: Не аутентифицирован / неверные креды / истёкший токен
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            invalidCredentials:
              value: { data: null, meta: { }, error: { code: AUTH_INVALID_CREDENTIALS, message: "Неверный email или пароль" } }
    ValidationFailed:
      description: Ошибка валидации полей запроса
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            invalidFields:
              value: { data: null, meta: { }, error: { code: VALIDATION_FAILED, message: "Ошибка валидации", details: { "email": "некорректный email" } } }
    Forbidden:
      description: Доступ запрещён (недостаточно прав или чужой ресурс)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            forbidden:
              value: { data: null, meta: { }, error: { code: ACCESS_DENIED, message: "Доступ запрещён" } }
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            notFound:
              value: { data: null, meta: { }, error: { code: NOT_FOUND, message: "Не найдено" } }
    Conflict:
      description: Конфликт уникальности (email/УНП уже заняты)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            duplicate:
              value: { data: null, meta: { }, error: { code: CONFLICT, message: "Пользователь с таким email уже существует" } }
...
  schemas:
...
    Envelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          nullable: true
          description: Полезная нагрузка; null при ошибке
        meta:
          $ref: "#/components/schemas/Meta"
        error:
          nullable: true
          $ref: "#/components/schemas/Error"
    Meta:
      type: object
      description: Для списков — пагинация; иначе пустой объект
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0, description: "Всего элементов" }
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Человекочитаемое сообщение (ru)
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: "Опционально: поля формы → текст ошибки, и т.п."
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { nullable: true, enum: [null] }
            error:
              $ref: "#/components/schemas/Error"

    # Каталог кодов ошибок error.code. Только расширяется; не переименовывается.
...
    CompanyStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    Company:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        shortName: { type: string, nullable: true }
        unp: { type: string, nullable: true, pattern: "^\\d{9}|\\d{10}$", description: "УНП РБ: 9 или 10 цифр" }
        legalAddress: { type: string, nullable: true }
        actualAddress: { type: string, nullable: true }
        bankAccount: { type: string, nullable: true }
        bankName: { type: string, nullable: true }
        bankBic: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        contactEmail: { type: string, format: email, nullable: true }
        status: { $ref: "#/components/schemas/CompanyStatus" }
        managerId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CompanyResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { $ref: "#/components/schemas/Company" }
    CompanyListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Company" }
...
    UpdateCompanyRequest:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        shortName: { type: string, maxLength: 255, nullable: true }
        unp: { type: string, pattern: "^\\d{9}|\\d{10}$", nullable: true }
        legalAddress: { type: string, maxLength: 500, nullable: true }
        actualAddress: { type: string, maxLength: 500, nullable: true }
        bankAccount: { type: string, maxLength: 50, nullable: true }
        bankName: { type: string, maxLength: 255, nullable: true }
        bankBic: { type: string, maxLength: 20, nullable: true }
        contactPhone: { type: string, maxLength: 50, nullable: true }
        contactEmail: { type: string, format: email, maxLength: 255, nullable: true }
        managerId: { type: string, format: uuid, nullable: true }
        status: { $ref: "#/components/schemas/CompanyStatus" }
...
    Category:
      type: object
      required: [id, name, slug, sort, isActive]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 255 }
        slug: { type: string, maxLength: 255, pattern: "^[a-z0-9-]+$" }
        parentId: { type: string, format: uuid, nullable: true }
        sort: { type: integer }
        isActive: { type: boolean }
        children:
          type: array
          items: { $ref: "#/components/schemas/Category" }
    CategoryRequest:
      type: object
      required: [name, slug, sort, active]
      properties:
        name: { type: string, maxLength: 255 }
        slug: { type: string, maxLength: 255, pattern: "^[a-z0-9-]+$" }
        parentId: { type: string, format: uuid, nullable: true }
        sort: { type: integer, default: 0 }
        active: { type: boolean, default: true }
    CategoryResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { $ref: "#/components/schemas/Category" }
    CategoryListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Category" }
    CategoryTreeResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Category" }
    Brand:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 255 }
        slug: { type: string, maxLength: 255, pattern: "^[a-z0-9-]+$" }
    BrandRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, maxLength: 255 }
        slug: { type: string, maxLength: 255, pattern: "^[a-z0-9-]+$" }
    BrandResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { $ref: "#/components/schemas/Brand" }
    BrandListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Brand" }
    Product:
      type: object
      required: [id, sku, name, categoryId, unit, vatRate, minStock, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        sku: { type: string, maxLength: 100 }
        name: { type: string, maxLength: 500 }
        categoryId: { type: string, format: uuid }
        brandId: { type: string, format: uuid, nullable: true }
        description: { type: string, nullable: true, maxLength: 4000 }
        unit: { type: string, maxLength: 20, default: "кг" }
        manufacturer: { type: string, nullable: true, maxLength: 255 }
        country: { type: string, nullable: true, maxLength: 100, default: "РБ" }
        tnvedCode: { type: string, nullable: true, maxLength: 20 }
        barcode: { type: string, nullable: true, maxLength: 50 }
        vatRate: { type: number, default: 10 }
        weightPerUnit: { type: number, nullable: true }
        minStock: { type: number, default: 0 }
        isActive: { type: boolean, default: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProductRequest:
      type: object
      required: [sku, name, categoryId, unit, vatRate, minStock, active]
      properties:
        sku: { type: string, maxLength: 100 }
        name: { type: string, maxLength: 500 }
        categoryId: { type: string, format: uuid }
        brandId: { type: string, format: uuid, nullable: true }
        description: { type: string, nullable: true, maxLength: 4000 }
        unit: { type: string, maxLength: 20, default: "кг" }
        manufacturer: { type: string, nullable: true, maxLength: 255 }
        country: { type: string, nullable: true, maxLength: 100, default: "РБ" }
        tnvedCode: { type: string, nullable: true, maxLength: 20 }
        barcode: { type: string, nullable: true, maxLength: 50 }
        vatRate: { type: number, default: 10 }
        weightPerUnit: { type: number, nullable: true }
        minStock: { type: number, default: 0 }
        active: { type: boolean, default: true }
    ProductResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { $ref: "#/components/schemas/Product" }
    ProductListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Product" }
    CatalogProduct:
      type: object
      required: [id, sku, name, unit, retailPrice]
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        name: { type: string }
        category: { $ref: "#/components/schemas/Category" }
        brand: { $ref: "#/components/schemas/Brand", nullable: true }
        unit: { type: string }
        description: { type: string, nullable: true }
        retailPrice: { type: number }
    CatalogProductResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data: { $ref: "#/components/schemas/CatalogProduct" }
    CatalogProductListResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/CatalogProduct" }